Upgrade center, zoom map params to floats

+++ malariagen_data/anoph/map_params.py
@@ -7,13 +7,13 @@ import xyzservices  # type: ignore
 from typing_extensions import Annotated, TypeAlias
 
 center: TypeAlias = Annotated[
-    Tuple[int, int],
+    Tuple[float, float],
     "Location to center the map.",
 ]
 
 center_default: center = (-2, 20)
 
-zoom: TypeAlias = Annotated[int, "Initial zoom level."]
+zoom: TypeAlias = Annotated[float, "Initial zoom level."]
 
 zoom_default: zoom = 3
